move parsers and scanners in the same directory
