--- src/common/type_mapper.py
@@ -1,37 +1,0 @@
-"""
-Type Mapper Module
-
-Centralizes logic for mapping external (Perl/Scanner) type strings to internal Python types.
-Ensures consistency between Scanner and Parser.
-"""
-
-from typing import Any, Dict
-
-
-class TypeMapper:
-    """
-    Static utility for mapping and normalizing stat types.
-    """
-
-    # Mapping from Perl/Legacy names to internal normalized names (lowercase usually)
-    # The registry uses lowercase names.
-    @classmethod
-    def normalize_type(cls, type_name: str) -> str:
-        """Normalize a type string (case-insensitive)."""
-        return type_name.lower()
-
-    @classmethod
-    def map_scan_result(cls, scan_result: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Normalize a dictionary result from the scanner.
-        Ensures 'type' field is consistent.
-        """
-        if "type" in scan_result:
-            scan_result["type"] = cls.normalize_type(scan_result["type"])
-        return scan_result
-
-    @staticmethod
-    def is_entry_type(type_name: str) -> bool:
-        """Check if type requires entries (Vector, Distribution, Histogram)."""
-        norm = type_name.lower()
-        return norm in ("vector", "distribution", "histogram")
