+++ src/core/__init__.py
@@ -1,0 +1,1 @@
+"""RING-5 Core Module - Provides infrastructure components."""
