+++ src/parsers/perl/libs/Scanning/Type/Configuration.pm
