+++ src/parsers/perl/libs/Scanning/Type/Scalar.pm
