+++ src/parsers/workers/__init__.py
@@ -7,8 +7,8 @@ Provides:
 - Gem5ParseWork: Worker for parsing gem5 stats files
 """
 
-from src.parsing.workers.gem5_parse_work import Gem5ParseWork
-from src.parsing.workers.parse_work import ParseWork
-from src.parsing.workers.pool import ParseWorkPool
+from src.parsers.workers.gem5_parse_work import Gem5ParseWork
+from src.parsers.workers.parse_work import ParseWork
+from src.parsers.workers.pool import ParseWorkPool
 
 __all__ = ["ParseWork", "Gem5ParseWork", "ParseWorkPool"]
