+++ tests/unit/test_data_parser_perl.py
@@ -2,8 +2,8 @@
 
 import pytest
 
-from src.common.types import StatTypeRegistry
-from src.parsing.parser import Gem5StatsParser
+from src.parsers.types import StatTypeRegistry
+from src.parsers.parser import Gem5StatsParser
 
 
 class TestGem5StatsParser:
