+++ tests/unit/test_histogram_rebinning.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.common.types.histogram import Histogram
+from src.parsers.types.histogram import Histogram
 
 
 def test_histogram_selection_priority():
