+++ tests/unit/test_scanning.py
@@ -2,8 +2,8 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from src.scanning.workers.gem5_scan_work import Gem5ScanWork
-from src.scanning.workers.pool import ScanWorkPool
+from src.parsers.workers.gem5_scan_work import Gem5ScanWork
+from src.parsers.workers.pool import ScanWorkPool
 
 
 def test_stats_scan_work_success():
@@ -11,7 +11,7 @@ def test_stats_scan_work_success():
 
     mock_vars = [{"name": "var1", "type": "scalar"}, {"name": "var2", "type": "vector"}]
 
-    with patch("src.scanning.scanner.Gem5StatsScanner.get_instance") as mock_scanner_cls:
+    with patch("src.parsers.scanner.Gem5StatsScanner.get_instance") as mock_scanner_cls:
         mock_instance = MagicMock()
         mock_scanner_cls.return_value = mock_instance
         mock_instance.scan_file.return_value = mock_vars
@@ -26,7 +26,7 @@ def test_stats_scan_work_success():
 def test_stats_scan_work_failure():
     work = Gem5ScanWork("test_file.txt")
 
-    with patch("src.scanning.scanner.Gem5StatsScanner.get_instance") as mock_scanner_cls:
+    with patch("src.parsers.scanner.Gem5StatsScanner.get_instance") as mock_scanner_cls:
         mock_instance = MagicMock()
         mock_scanner_cls.return_value = mock_instance
         mock_instance.scan_file.side_effect = Exception("Scan error")
